+++ src/main/java/com/capstone/goat/service/GameService.java
@@ -110,8 +110,8 @@ public class GameService {
                 .orElseThrow(() -> new CustomException(CustomErrorCode.GAME_NOT_FOUND));
 
         PreferCourt preferCourt = preferCourtRepository.findFirstByCourtAndGameId(court, gameId);
-        float latitude = preferCourt.getLatitude();
-        float longitude = preferCourt.getLongitude();
+        double latitude = preferCourt.getLatitude();
+        double longitude = preferCourt.getLongitude();
         game.determineCourt(court, latitude, longitude);
     }
 
@@ -201,8 +201,8 @@ public class GameService {
         }
 
         PreferCourt preferCourt = preferCourtRepository.findFirstByCourtAndGameId(courtName, gameId);
-        float latitude = preferCourt.getLatitude();
-        float longitude = preferCourt.getLongitude();
+        double latitude = preferCourt.getLatitude();
+        double longitude = preferCourt.getLongitude();
         game.determineCourt(courtName, latitude, longitude);
     }
